Update patch set 1
Patch Set 1:

"The 'rockbox way' is to turn off backlight when hold is engaged."

Still the same with this patch. With the added option of toggling the backlight/touchscreen with the power button while in the WPS.

"Describing WPS as 'main screen' is questionable [...]"

If the WPS is not the main screen of a software player than what screen is?

Patch-set: 1