Update patch set 4
Patch Set 4:

No, I will have to the modification Thomas wants, if I cannot convince him to leave it as it is right now. At the moment he wants me to set ROCKBOX_DIR to "/mnt/sdcard/.rockbox" but when I tried his patch it did not work and I will have to wait until this is fixed.

Patch-set: 4