Update patch set 1
Patch Set 1:

My need is often that i want to know "what song is this that is playing" or "how much battery is left". So when I want to do this without my patch I can either turn on the option "turn on backlight on 1st keypress". This has the disadvantage that when I want to skip a song, I have to press the button twice. Or I can leave the seting disabled, and press Vol- to turn on screen and then Vol+ to set back the volume.
But probably the main problem is that DX50 has a very sensitive touchscreen, that is why DX50 users prefer this mod: When screen goes "on", the touchscreen is activated. When I want to skip a track with the player in my pocket, the file gets deleted because I did some touching of the touchscreen that I did not want.

Patch-set: 1