Update patch set 1
Patch Set 1:

Doh!
I thought it was so obvious that I never explicitly named the touchscreen as the problem here:
http://forums.rockbox.org/index.php/topic,48058.msg229859.html#msg229859
Sorry for that.

So to reiterate: The problerm is not the backlight. The problem is the touchscreen coupled to the backlight.
This makes blind usage (e.g. in your pocket) very annyoing because when the backlight turns on the touchscreen turns on. And it is a very sensitiv touchscreen, very easy to activate functions you do not want at that particular moment.

Patch-set: 1