Update patch set 1
Patch Set 1:

I did some mor thinking about this patch and I got to the conclusion that this mod is very good as it is for touchscreen targets and some people might even prefer this mod for newer non-touchscreen targets that have a power button. I don't remember why I did the cancel action from main menu not include I'm the keymap. But I think it was not possible. I will look at this and look on what targets this mod makes sense and prepare them so that only the preprocessor flag has to be set to enable it. I don't give up with this because as I already said, DX50/DX90 users need this or a similar mod.

Patch-set: 1