Update patch set 1
Patch Set 1:

Instead of modifying the cancel action in the main menu I will add a new action to resume playback that can be called used with keymaps. This action does not exist yet (that is why i did not include it in the keymap). The whole modification I will prepare to make it easily appliable for targets cowond2 and creativezanxfi2 besides ibassodx50 and ibassodx90 which are already prepared.

Patch-set: 1
Label: Code-Review=0