Update patch set 2
Patch Set 2:

New Patch set: I changed the commit message! There are only few touchscreen targets that can profit from this patch. On creative zen Xfi2 it does not make sense so I did not prepare the patch for this target. New is: Now the cancel action is done in the keymap file (thanks Marcin, you were correct, there is an action for this, but not all keymaps were up to date to use main menu keys). Cowon D2 target is now also prepared but the define MODERN_UI is commented out for it. On Cowon D2 it also does not make as much sense as it does on DX50/DX90. The patch seems to be quite a unique need for this targets until now. It is the combination of having Touchscreen AND buttons that are interesting to use blind when player is in a pocket (play,pause,skip,volume). But every future target that has this combination will benefit from this patch. Maybe the name "Modern UI" is an unfortunate choice; better suggestions are welcome :-)

Patch-set: 2