Update patch set 2
Patch Set 2:

I was not aware that there are such devices, but it would not make sense for them. Which ones have translucent color screen? To be honest, I was very surprised that there is no define for displays that are passively readable. I thought that this difference is an important thing that would make more code differences necessary than the ones we have.

Patch-set: 2