Update patch set 2
Patch Set 2:

I knew that if there are reasons against they will pop up and rockbox will definitely not ignore a blind minority for a smaller menu. Thanks Avi. So at boot, if the setting "Reset Backlight Off at Boot" is set, it will take the default value if the "Backlight Timeout" is set to "-1". To be honest i don't want to make the menu even more complex with such settings. I thought that the backlight timeout settings was something that nobody looked at after all those Color LCD came up. I think this is a patch on its way to be abandoned. One other Idea I had is to make a compile option for blind users to optimize both builds, this would maybe allow to leave the "off" option away. I also think that rockbox will become modular in the future, so maybe it is not yet time for this.

Patch-set: 2