Update patch set 2
Patch Set 2:

The new define i would name HAVE_PASSIVELY_READABLE_SCREEN or similar, because I think most of the future targets will not be passively readable so it would be the regular case.
AT the moment i think that every target handles its LCD off in the backlight-xxx.c and the developer can decidecide if LCD has to go off together with the backlight if the device allows to handle this things seperatly. We went a bit off topic but i think Avi is right that we cannot take out an option that someone possibly could like. The downside of leaving it aß it is, is that unexperienced users who accidentally switched off the backlight forever start to hate rockbox.

Patch-set: 2
Reviewer: Gerrit User 1000386 <1000386@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0