Rockbox system settings for iBasso DX50: USB mass storage, USB charge only, adb over USB.
This patch adds a new setting in Settings -> General -> System: USB Mode.

Possible settings are:
- Mass Storage: The default, on USB connect export the internal and external
drives as usual. This will crash Rockbox when it is running from the internal
drive (/mnt/sdcard) and not the system partition.
- Charge Only: USB charge only, do not export the internal and external drives.
This will not crash Rockbox.
- ADB: Enable the Android Debug Bridge. This will not crash Rockbox.

Charge only and ADB are actually the same, since it is not yet established if
charge only is doable without adb and major hooks into Android.

German translation provided.

Change-Id: I616247c29946c6595ebcf8f0c14b2410c9f0498b