Update patch set 2
Patch Set 2:

I' m very happy about this patch and can't wait to try it. For the non-DXx0 owners: until now, the USB mode was set permanrntly by the official firmware. So if we wanted to enable adb, we had to reboot to OF to change the mode and reboot to OF. This took quite a while.
What I can say until now is that you should add the header for the new files and that I would name the settings android instead of dx50 for the language system but I'm not sure if this would really be better. Thank you very much for the patch, it will be very helpful, specially at current state where we have to try out many things. One question: USB DAC mode is available in the OF setting, do you know what this does?

Patch-set: 2