Update patch set 3
Patch Set 3:

On the fuze+ the playability crumbles to zero the first time you need to duck and crawl, this is caused by the fuze+'s odd touchpad input.  It has enough buttons, but they are virtual spaces on the touch pad that can only be activated one at a time (no multi touch support).

Once I tried to quickly add down_right, down_left, up_right and up_left to the xrick sources, but failed miserably (I was ANDing the inputs together).  I'm not as great with C as I wish I was.  support for the diagonals would help this port reach a wider range of targets.  Xrick would an excellent addition to rockbox IMO.

Patch-set: 3