Update patch set 2
Patch Set 2: Code-Review+1 Tested+1

I've tested this patch on my DX50 and it works very well. This patch is incredibly helpful! One little thing I've noticed is that if I save all settings to a cfg file, change the mode and reload the configuration file, the USB mode is not switched back. But this is a debug setting anyway, in my opinion not necessary to change. I'm very happy with the patch in general, the position in the setting menu is perfect. IMHO maybe the definition for the lang handling could maybe be a bit more general (ANDROID_USB_MODE or USB_MODE instead of DX50_USB_MODE). DX90: I think the target DX90 could wait this patch (as you did) and when the DX90 developer (alias headwhacker) comes here (he told me that this are his plans), he can enable it on the DX90 if he sees that it is working for it too.

Patch-set: 2
Label: Code-Review=+1
Label: Tested=+1