YH8xx,YH9xx: Keymap improvement (WIP)
The main innovation in this patch are two "virtual buttons"
for the record switch on YH92x targets. When the switch state
changes, a single BUTTON_REC_SW_ON or .._OFF button event will
be generated. Thus keymap code can react on switching, but
not on the actual state of the switch.

NOTE: don't push yet, this patch will break most plugins
on YH92x! Reworked Plugin keymaps will be added bit for bit.

Change-Id: Ic8d1db9cc6869daed8dda98990dfdf7f6fd5d5a1