Update patch set 3
Patch Set 3:

The crashing has nothing to with this patch, it is the same as ever with DX50 and mass_storage. The screen just freezes, no error output. I did not check the log messages yet. I think it happens as soon as /mnt/sdcard is unmounted. In the future (patch 993) we will even run rockbox from the unmounted partition.

Patch-set: 3
Label: Code-Review=0