Update patch set 1
Patch Set 1:

This patch solves some extrange errors that appeared while trying to run Doom in SVN-21141108 (logf enabled, gcc 4.4.4, ld 2.20.120100303). It is not easy to notice the existence of this problem, the sound was apparently good for me, USB transfers were also correct, as well as video playback, no problems reinitilizaing the database... Finally i found that test_codecs, imageviewer and test_mem_jpeg are also failing, some failures only occurs when the USB is connected.

After debugging, it seems that due the misalignment of the ATA storage buffer used for DMA transfers, occasionally corrupted data is read. 
   
Although this patch affect other functionalities, mainly audio buffering, i think it should not be harmful and must be commited as soon as possible. The only reason for not doing so is that it was a quick patch and has only been tested for a couple of hours, i need a little more time to revise it and confirm there are no side effects.

Patch-set: 1