Update patch set 1
Patch Set 1:

Yes, it can be verified in debug screen, it is an arm926ej-s with 16+16 Kb cache and 32 bytes cache line. In addition while debugging i observed several readed HDD sectors with the first 16 bytes corrupted. It seems it is due the buffer was aligned to 16 (not 32) and the first 16 bytes for the 32-byte line are assigned in .bss to other related variables, this buffer is discarded from d-cache to read DMA data, then anything can happen.

Patch-set: 1
Reviewer: Gerrit User 1000097 <1000097@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0