Update patch set 2
Patch Set 2:

Not if you think, that an empty header to satisfy include dependencies is a good idea.

From my point of view someone forgot proper #ifdefs around this include (probaly ! (CONFIG_PLATFORM & PLATFORM_HOSTED)).

I thought the safest way was to just ignore this include for these devices and wait for the experts to decide what to do with this.

Patch-set: 2
Label: Code-Review=0