Revert "Rewrite filesystem code (WIP)"
This reverts commit 7d1a47cf13726c95ac46027156cc12dd9da5b855.

Commit 7d1a47cf ("Rewrite filesystem code (WIP)") introduced
regressions ([1],[2]) on several rk27xx targets, making them nearly unusable.

Due to size of this change it's nearly impossible to find the root
of the problem, so I suggest to revert it.

I'm going to investigate this further, but this definitely will take a lot of
time. I think reverting this until we'll find WTH is happening is the best
option here.

[1] http://forums.rockbox.org/index.php/topic,48658.0.html
[2] http://forums.rockbox.org/index.php/topic,48675.msg230556.html#msg230556

Change-Id: Ib4fab6cacadb6f76dbd6a9331cd5f3f786fb5ba7