Update patch set 2
Patch Set 2:

A fallback aligned_buf is not thread-safe. I'd rather fall back to the for-loop instead.

Yes, multithreading is not an issue now, but IMO the driver should give control to other threads during  the DMA transfer instead of hogging the CPU until the transfer is done. Then the driver must be MT-safe.

Patch-set: 2