Update patch set 5
Patch Set 5:

I probably won't have time to work on this for a few days, but I am ok with splitting the codec changes.  If no one beats me to it, I will look at it eventually.  

Before this can be committed I'd like to figure out what the issues are with the Clipv2 higher PCLK (Bertrik may have some ideas) and also enable it for the Fuze V2 (which will require someone else to test as I don't have a FuzeV2).

Patch-set: 5