Update patch set 3
Patch Set 3:

I see...Still, it would be nicer if it were dynamic. i.e. proposing a list of *existing* governors, that can be easily fetched from the sysfs interface.
Would you mind trying to make it generic for the hosted platform? And preferably in a separate patch (if the governor "driver" was not part of the cleanup patch)

Patch-set: 3