Update patch set 8
Patch Set 8:

Sorry, should have been more clear:  I don't expect you to do this now, just suggesting that the comment could be a little more clear.  It can be worked on in the future if someone is interested.  

As for the ADC changes, sorry to ask twice.  That seems fine to me.

Patch-set: 8
Label: Code-Review=0