Update patch set 10
Patch Set 10:

With this change all buttons are ignored when backlight is off exept the predefined ones (normally: power button). If ignore first keypress would still be enabled, you would have to press power button twice to wake up screen instead of only once. That would be the only difference because the screen could not be waken up in a different way. For the YP-R1, I would be honoured if you would try it out. You will have to enable it in samsungypr1.h and add the power button to button-target.h

Patch-set: 10