Update patch set 8
Patch Set 8:

Forgot to mention that the run-time math is now in fixed point and thus the logf code copy is removed.

The accuracy of the metronome is determined by timer drift (still don't quite get why it is so strong when using smaller intervals ... what should be safe?) and the difference between PLLs. The scheduler doesn't seem to have a big hit on that on my Sansa Clip+. When recording click tracks to my laptop, I only see jumps from the quantization on timer intervals.

Patch-set: 8