Update patch set 11
Patch Set 11: Tested+1

I see that I can give +1 for it being tested, which it is on Sansa Clip+ hardware and simulator as I indicated previously.

Should I also mark the "+1 looks good to me" to indicate that I as author want this to be merged or is this exclusively for other reviewers? Because ... I don't have anything to add right now.

Patch-set: 11
Label: Tested=+1