Update patch set 11
Patch Set 11:

Does tap-tempo work with the old version of the plugin? What kind of "weird" changes are that? The logic just counts the time since the last tap and sets the tempo from that (no running average or such).

I didn't change the basic logic, but the timer setup is sligtly changed.

Patch-set: 11