Update patch set 12
Patch Set 12:

Thanks for elaborating. I fixed the doubled tempo, a bit embarassed to see that it is not model-specific and I failed to see it on the Clip+. It was related to the changed timer frequency division and the hardcoded factor in the tap tempo calculation.

About the annoying long pause after tapping: Yes, I had that eliminated at first, but that causes irrigating double ticks if you intend to slow down the tempo via tapping. It's a separate issue to find nice mechanics for tapping that feels natural.

About the feature of tapping while paused: I changed it now so that tapping starts metronome playback. That way, the timer can still be disabled for paused metronome (and save some power). So, the behavioural difference to the old metronome now is that it automatically starts when you tap, instead of just recording the tempo change and waiting for you to hold select. Is that acceptable?

Patch-set: 12