Update patch set 12
Patch Set 12:

Can you test if the double beats also vanish if you set timerfreq_div=1024 (the old value) or some other high value that actually is a better divider for the timer freqency (which is not 1024-based, I presume)? Because the actual timer mechanism is unchanged. It works with >500 also for the Sansa Clip+, it's just that it doesn't really improve things and rather introduces more drift (slightly slower bpm than configured).

Although ... Off-hand I'd assume that the smaller divisor (larger time interval) is more easy on the CPU ... the real change from the old plugin is not the timer mechanism but the display indication. So, you could try to deactivate the trigger_display() calls in play_ticktock() and see if that changes things for some reason.

So far I don't yet have an idea how an accidental double-tick (or tock, in the classic mode;-) comes to be.

Patch-set: 12