Update patch set 12
Patch Set 12:

Interesting. And sad. I don't see anything wrong with the display drawing (as it is _not_ happening in the timer callback), but I guess this means your CPU is too slow to keep on playing the sound while drawing the display. Could your double-tick come to be by the audio "thread" (or whatever) becoming interrupted by the display drawing?

So, my other theory is now to do the drawing _before_ playing the tick. That might work, but it will also introduce delay. But if that is always the same delay, it won't matter much to the metronome. Can I upload an experimental patch for testing? I don't want to push another patch set yet, but would like you to test some tweaks that could be part of patch set 13.

Patch-set: 12
Label: Code-Review=0