Update patch set 13
Patch Set 13:

What the heck ... I modified the patch to draw the display right before starting playback now, as it has been on the first iterations, actually (although inside the timer callback).

Does it still do double-ticks on the YH820? If yes, we might just have to put #ifdef in to skip the display drawing on that hardware.

Patch-set: 13