Update patch set 13
Patch Set 13:

OK, then. Performance issues at tempos approaching 400 bpm are perhaps not too surprising. But since the display becomes useless anyway, I could introduce a bpm limit (actually: limit of beat duration, as it could be a bout 1/8 beats and the bpm number is always about 1/4) above which the display is not drawn, but I guess we agree that that's not worth it.

I assume musicians also experience performance problems with 400 bpm ticks;-)

What I need to do to cleanup the current state is to fix the situation for HWCODEC again. Currently, there is display drawing happening in the timer callback for those. I'll fix that and resubmit the final patch, for which I ask you to provide the "tested" flag again. I'll throw in a flag after my own final tests, too, but I guess "code-review" needs really a different person than the one who wrote it.

Patch-set: 13
Label: Code-Review=0