iBasso DX50/DX90: Rockbox loader, replacement for original firmware MangoPlayer.
This loader is needed to use Rockbox on iBasso DX50/DX90 devices. It replaces
/system/bin/MangoPlayer.

It expects
- the original firmware MangoPlayer as /system/bin/MangoPlayer_original
- the chooser.bmp file as /system/chooser.bmp
- the rbmissing.bmp file as /system/rbmissing.bmp
- the usb.bmp file as  /system/usb.bmp
- Rockbox in /mnt/sdcard/.rockbox/ (internal device storage accessable via USB)

On first boot or if the device is booted with the hold switch engaged (upper, locked position) this
loader will let the user choose between Rockbox and the orginal firmware MangoPlayer. It remembers
the last choice and will start the choosen player an following boots.

In the chooser screen
- the device can be powered off with a power button long press.
- Rockbox can be startet with the "Next/Fast Forward" button or by touching the right part of the
touchscreen.
- the orginal firmware MangoPlayer can be startet with the "Previous/Rewind" button or by touching
the left part of the touchscreen.

If Rockbox can not be found/executed than a "Rockbox is missing" screen will be shown and the
orginal firmware MangoPlayer will be started after 30 seconds.

This loader will (try to) restart Rockbox or the orginal firmware MangoPlayer if something
unexpected happens.

This loader can be installed via a customized firmware image (update.img) or manually via the Android
Debug Bridge (if you get adbd started on the device).

This is a rewrite and code clean up of the original submission, eliminating the need for a custom
Android Vold. This loader "knows", when Rockbox has exited due to a USB mass storage connection
remounting /mnt/sdcard and acts accordingly, displaying a "USB connection" screen while waiting to
restart Rockbox.

Change-Id: I2698e173437f9c7aa1ff40649a290220e2ee0439