Update patch set 14
Patch Set 14: Tested+1

Well, then, here it is: The hopefully final version of the enhanced metronome patch. I corrected the HWCODEC case again to not do display drawing in the timer callback.

I tested on the Sansa Clip+ and it does 1 to 400 bpm nicely in the simple mode. It also handles beat patterns using 16th notes at 120 bpm, so effectively 480 tick/tocks per minute.

An imaginable enhancement for some machines may be to only draw some pixels on the corner of the display instead of filling rectancles, but I don't know what the actual constraints of those display controllers are.

Anyhow, I hereby ask for inclusion of that version of the patch as an all-around improvement over the old metronome.

Patch-set: 14