Update patch set 1
Patch Set 1:

I'm unsure if buflib is a great fit for codecs. I guess it depends on the codec. But it should never use the core context, because that means the core potentially (negatively) impacts codec operation.

This is the main reason that the codec buffer is still static and not allocated via buflib. Reusing the core's context would make it even worse.

Patch-set: 1
Reviewer: Gerrit User 1000004 <1000004@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0