Update patch set 11
Patch Set 11:

Thinking about this more carefully, I guess software volume control operates on 16 bit samples because it has to be applied to  audio that has already been placed in the PCM buffer (otherwise there would be a very long lag when changing volume), so it must be done on 16 bit (or N bit) audio.  So you would have to change that function too to work on 24 bit audio, or ideally not use software volume control if there is a hardware option (i noticed PLATFORM_HAS_VOLUME_CHANGE is commented out in the DX50 config file).

Patch-set: 11
Label: Code-Review=0