Update patch set 14
Patch Set 14: -Tested

OK, we're getting to something. You're right that after a certain tempo, the visual indication doesn't help humans anymore. A test case that's still reasonable on the Clip+, though, is playing this tempomap:

funky: 4 16/16 120 x.x.X..X.Xx.X..X

Most of the beats are spaced a 1/8 notes (240 actual bpm), while once you got a 480 bpm interval, which still makes sence visually, as it's a switch from tick to tock and you can see that on the display.

But on the slow machine, this would already pose a problem, I presume.

Can you confirm what exactly happens when you delete lines 1359 to 1361? I mean disabling the /* Clear blinker. */ part totally. Does that still make things out of sync?

If it does not, I can easily introduce some means to avoid the blinker clearing for faster tempi, as it's not noticeable anyway. I just have it in there for slow tempo to give a visual hint on when the next beat will occur. It really helps anticipation. At small beat intervals, it can be deactivated anyway.

The alternative of detecting the machine being slow can be considered, but I'm not sure how good a mechanism I introduce will be. I'd need a device to test. I wonder if I can underclock my Sansa Clip+ to simulate things.

Patch-set: 14
Label: Tested=0