Update patch set 15
Patch Set 15: Code-Review+1

(3 comments)

This looks a lot better and is so much nicer to read.

A few minor things, none of which should be interpreted as a blocker:
* inline comments
* lcd-ibasso.c looks similar to the ypr0 version, merging them should be looked at
* ibasso has its own pcm-alsa driver, while there is also a (supposedly) generic one. I don't know if tinyalsa is different enough, but if not these could potentially also merged (doesnt need to happen now)

The rest looks great and is mostly self-contained for the ibasso devices.

Patch-set: 15
Reviewer: Gerrit User 1000004 <1000004@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=+1