Update patch set 10
Patch Set 10: Code-Review-2

This kind of change has been requested a lot of time, also on non-touchscreen devices, so I think we should come up with a solution. 

But this is in such a high level/user intarface layer that we really want to have a generic solution, and not one that creates two families with a MODERN_UI define (that name is kind of questionable too).

Patch-set: 10
Reviewer: Gerrit User 1000004 <1000004@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=-2