Update patch set 18
Patch Set 18:

It's getting early in the morning. So this is a state that might or might not be "final" again. I would like to be done with this, but, well, it seems that it has to be perfect (for some definiton of perfect).

After it has been suggested on IRC to test various target builds, I wondered about HWCODEC again and enabled the metronome for the simulator build (disabling actual MP3 playback). There's some understated visual indication now, too. Main change in logic is that it works much more like SWCODEC now, with playback and drawing happening in the main loop.

Works fine in sim, but is there someone out there who actually can test this on an Archos device with 2-line LCD?

Patch-set: 18