Update patch set 18
Patch Set 18: Code-Review+1 Tested+1

Well, then. I tested the build of all targets mentioned on that automated build site that got passed to me via IRC, with simulator and manual builds in addition to plain builds and only those fail:

android480x800
android480x800mips
ibassodx50
iriverh120htmlmanual
samsungypr0
samsungypr1

Reason: I don't have Android SDK and also the toolchain for samsungypr is bitchy (doesn't want to install into a prefix that is user-writable but that the user is not allowed to delete, i.e. a directory under /opt). The manual build has some error from pdftex.

I don't see any trouble associated with the metronome there and it still works on my Clip+ as well as the simulator of an Archos Player.

If there would be someone able to test the metronome on an actual Archos Player device (anything with HWCODEC and/or a non-bitmap LCD), that would be nice as I can only guess that the HWCODEC stuff works, especially since I unified the playback always happening in the main thread instead of the callback.

Are there Archos devices still alive?

Patch-set: 18
Label: Code-Review=+1
Label: Tested=+1