Update patch set 10
Patch Set 10:

IMHO the way the "lock" works should be different depending on the physics of every target. On DX50 the hold switch should really do a lock because it is a physical slide switch. But this configurable partial lock is a very good idea because the device has buttons that are used for playback control: Play/Pause, Next/FFW, Prev/RRW, Vol+, Vol-. Cowon and FIIO have often such settings in the OF and many people use them. On many players this change would not make sense of course. It is the simple fact that parts of the buttons can be used for playback control. Something that is hard to implement but would make this mod even better is that when lock mode is entered, the player should jump from wherever it is at the moment to WPS.

Patch-set: 10
Label: Code-Review=0