Update patch set 4
Patch Set 4:

I guess it's worth restating here that it looks like when the transfer queue is seen as empty and because interrupts are stacked, it's possible to trigger a restart on the interface before all the data is clocked out during a write. Interrupts also have to be masked during the ISR in certain spots to protect list integrity because of stacking. With the I2C hardware somewhat undocumented I have to wonder if there aren't any more useful status bits and interrupts available to help. If not, I'll just attempt to make due.

Patch-set: 4