Make rockbox for Android work on ART runtime
This is *HUGE* hack.
The main problem is our threading. ART introduced stack overflow
check which we trigger if JNI calls are not from main thread.
Hence button_queue is abused to pass function pointer which
is going to be called from main thread. Problematic JNI
calls are abstracted as void(void) functions which are pushed
to button_queue() and are executed in main thread.

Other then this there are a few places where NewGlobalRef() was
needed.

I hacked configure slightly to build with latest NDK and SDK.

Change-Id: I9675c1b90296ab26715db96066bdf52e33e166c5