[WIP] DuckyScript plugin
Allows a DAP to behave like a USB Rubber Ducky by executing DuckyScript files,
which is a BASIC-like language with facilities for sending keystrokes.

This plugin also implements several extensions to vanilla DuckyScript, all of
which are documented with comments in the source.

Also changes some entries in the HID usage tables.

To test, type this in a file named `test.ds`, run it, and plug in USB:

STRING The quick brown fox jumps over the lazy dog.

TODO:

 - Manual entry
 - Test compatibility with the original DuckyScript
 - Test language features (variables, control flow, expression parsing)
 - FASTER TYPING !!!
   - Currently averages ~96 keys/sec.
   - Theoretical maximum: 500 keys/sec.
   - Of interest: G#1223 (http://gerrit.rockbox.org/r/1223/)

Change-Id: I149ecd08ba124f8b8637a002ee2652dd54a44e58