[WIP] DuckyScript plugin
Allows a DAP to behave like a USB Rubber Ducky by executing DuckyScript files.
This plugin also implements several extensions to vanilla DuckyScript, all of
which are documented in the source.

To test, type this in a file named `test.ds`, run it, and plug in USB:

STRING The quick brown fox jumps over the lazy dog.

NOTE: Also changes some entries in the HID usage tables.

TODO:

 - Manual entry
 - Test compatibility with the original DuckyScript
 - Test language features (variables, control flow, expression parsing)
 - FASTER TYPING !!!
   - Currently averages ~96 keys/sec.
   - The bottleneck is having some USB code in a thread, which is bounded by HZ
   - Theoretical maximum: 500 keys/sec.

Change-Id: I149ecd08ba124f8b8637a002ee2652dd54a44e58