iPod Classic: dual bootloader - v1
First of all, this bootloader is based at 90% on emCORE code,
so you know who to blame if you iPod gets bricked.

Seriously, it is very difficult (if not impossible) to brick an
iPod 6G. At most it could need to be restored using iTunes, so
backup your songs before installing the dualboot, anyways this
should not happen with this current version, i have tested it
dozens of times on two devices with no problem.

mkdfu generates a .dfu file containing an installer/uninstaller
and the RB bootloader. The installer moves the original NORBOOT
and install the RB dual bootloader. When the iPod boots then the
RB dual bootloader launch Rockbox or the OF using the original
NORBOOT.

See mk6gboot/README for more info on how to install.

I started this as a preliminary version, actually it is functional,
but lots of things needs to be done:

TODO:
- mk6gboot: only tested on Linux, needs to integrate ipoddfu_c
  utility by user890104, add the new target to RB installation
  utility, and test everything on Windows and Linux.
- some rearrangement of the code, there is a clickwheel patch
  included that should be separated, optimize NOR read/write
  using DMA, battery trap...
- DFU graphic install/uninstall, does it worth?
- ...

Patch Set 2:

- added hash for firmware 2.0.1 (thanks to user890104)
- added notes on how the whole thing works (see README.txt,
  dualboot.c and bootloader/ipod6g.c)

Note that you need g#1219 and g#1220 to compile the bootloader,
a patched version of RB (rockbox.zip) should be installed into
the iPod partition (FAT).

Patch Set 3:

- added menu to choose a default FW to load, default options are
  saved into a PMU register, press SELECT+RIGTH at boot time to
  enter the options menu.
- added battery trap, it is a colored text line.

TODO:
- battery readings differ too much when USB (or wall adaptor) is
  plugged/unplugged, need to find a way to deal with this issue
  in RB/bootloader.
- RB deserves some sort of decent graphic progress for the battery
  trap, it will be seen so often.

Patch Set 4:

- added hash for firmware 2.0.5 (not tested).
- boot options menu is hidden, it is shown when emCORE binary is
  present, otherwise SELECT+RIGHT silently switches the default FW.
- updated mk6gboot.exe utility, integrates ipoddfu to scan for USB
  DFU devices and send .dfu files, supports Apple drivers (using
  winapi) and libusb drivers, see mk6gboot/README.
- usable RockboxUtility.exe to install/uninstall the bootloader
  (Windows and Linux), there are a few issues:

  It is patched to use local files (not downloaded from the web),
  bootloader-ipod6g.ipod and rockbox.zip must be located in the
  same place as RockboxUtility.exe

  Other versions of Rockbox (i.e. last git versions) are not
  compatible with the bootloader (should be fixed soon), you must
  use the rockbox.zip built using the same patch set as the
  bootloader. Rockbox version showed by RockboxUtility is wrong
  and should be ignored.

  At this moment, bootloader installation status cannot be checked,
  for RockboxUtility it is always detected as installed, even if
  there is no bootloader installed, ignore this.

  While installing/uninstalling the bootloader, do not close the
  progress dialog using the close window button (top-right), use
  the 'Abort' button to do that. If the progress window is closed
  using top-right [X] button then the install/uninstall process
  continues on the background and anything could happen, close the
  whole RockboxUtility and relaunch it again.

Change-Id: If45e2927ee75ae797efbbf5675f2b454fdc72f5d