iPod Classic: dual bootloader - v1
First of all, this bootloader is based at 90% on emCORE code,
so you know who to blame if you iPod gets bricked.

Seriously, it is very difficult (if not impossible) to brick an
iPod 6G. At most it could need to be restored using iTunes, so
backup your songs before installing the dualboot, anyways this
should not happen with this current version, i have tested it
dozens of times on two devices with no problem.

mkdfu generates a .dfu file containing an installer/uninstaller
and the RB bootloader. The installer moves the original NORBOOT
and install the RB dual bootloader. When the iPod boots then the
RB dual bootloader launch Rockbox or the OF using the original
NORBOOT.

See mk6gboot/README for more info on how to install.

I started this as a preliminary version, actually it is functional,
but lots of things needs to be done:

TODO:
- mk6gboot: only tested on Linux, needs to integrate ipoddfu_c
  utility by user890104, add the new target to RB installation
  utility, and test everything on Windows and Linux.
- some rearrangement of the code, there is a clickwheel patch
  included that should be separated, optimize NOR read/write
  using DMA, battery trap...
- DFU graphic install/uninstall, does it worth?
- ...

Patch Set 2:

- added hash for firmware 2.0.1 (thanks to user890104)
- added notes on how the whole thing works (see README.txt,
  dualboot.c and bootloader/ipod6g.c)

Note that you need g#1219 and g#1220 to compile the bootloader,
a patched version of RB (rockbox.zip) should be installed into
the iPod partition (FAT).

Change-Id: If45e2927ee75ae797efbbf5675f2b454fdc72f5d