An almost working .gob file viewer
A viewer plugin that displays a list of the files in a Dark Forces .gob file.
It very often crashes due to the fact that it declares a function within another function.
This could be semi-trivially fixed but it will not be because this commit is now ABANDONED.

All work for this will now be devoted to a full fledged Dark Forces Game.

Change-Id: Icf729e3f27ada25949b86932ab828abfcc7a223c