hwstub: rewrite and expand library
Rewrite the hwstub library in C++, with a clean and modular design.
The library was designed from the ground up to be aware of multithreading
issue and to handle memory allocation nicely with shared pointers.

Compared to the original library, it brings the following major features:
- support for JZ boot devices, it is very easy to add support for others
- support for network transparent operations (through sockets)

Also port all the existing tools to the new library.

Note: this is a WIP.
The following have been done:
- implement all USB part of the library
- port hwstub_load, hwstub_test and hwstub_shell to use the new library
Still to be done:
- implement TCP/net/server part of the library (WIP)
- port hwstub_server to use the new library (WIP)
- implement a context that forks as a server in the library

Change-Id: I75899cb9c7aa938c17ede2bb3f468e7a55d625b4